requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
numpy>=1.23.0
python-dotenv>=1.0.0